Updated CPO to be compliant with JDBC 4.3 - Added new JDBC Types - Added Java Types that JDBC will convert using setObject(index) and   getObject(index, Class<T>) - Modified the ResultSet getters, PreparedStatementSetters and   CallStatement getters and setters to use the new setObject and getObject facilities - Added a new method mapper type for the setObject and getObject   facilities - Changed the exporter code to use the javaClass and not the dataClass   as they can be different now. - Added a test to test the setObject and getObject

+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/ValueObjectBean.java
@@ -21,6 +21,7 @@
 package org.synchronoss.cpo.jdbc;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.*;
 
 /**
@@ -53,7 +54,7 @@ public class ValueObjectBean implements ValueObject, java.io.Serializable, Clone
   private boolean attrBit_ = false;
   private BigDecimal attrTinyInt_ = null;
   private int attrSmallInt_ = 0;
-  private BigDecimal attrBigInt_ = null;
+  private BigInteger attrBigInt_ = null;
   private BigDecimal attrReal_ = null;
   private byte[] attrBinary_ = null;
   private byte[] attrVarBinary_ = null;
@@ -149,7 +150,7 @@ public class ValueObjectBean implements ValueObject, java.io.Serializable, Clone
     return attrSmallInt_;
   }
 
-  public BigDecimal getAttrBigInt() {
+  public BigInteger getAttrBigInt() {
     return attrBigInt_;
   }
 
@@ -249,7 +250,7 @@ public class ValueObjectBean implements ValueObject, java.io.Serializable, Clone
     attrSmallInt_ = attrSmallInt;
   }
 
-  public void setAttrBigInt(BigDecimal attrBigInt) {
+  public void setAttrBigInt(BigInteger attrBigInt) {
     attrBigInt_ = attrBigInt;
   }
